requirements.txt
# Crosshatch - Modern Implementation
# Python 3.10+ required

# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0

# PyTorch (CPU version - for GPU, see installation notes below)
torch>=2.5.1
torchvision>=0.20.1

# For Docker API
Flask>=3.0.0
Werkzeug>=3.0.0

# For notebooks (optional)
jupyter>=1.0.0
matplotlib>=3.7.0
ipython>=8.12.0

# SAM2 - Install separately:
# git clone https://github.com/facebookresearch/sam2.git
# cd sam2
# pip install -e .
# ./checkpoints/download_ckpts.sh

# GPU Installation Notes:
# For CUDA support, install PyTorch with CUDA:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# (adjust cu121 to your CUDA version)
